+++ packages/plugin-packages/mashroom-messaging-external-provider-amqp/src/topic_converter.ts
@@ -1,0 +1,5 @@
+
+export const topicToRoutingKey = (topic: string): string => topic.replace(/\//g, '.');
+
+export const routingKeyToTopic = (routingKey: string): string => routingKey.replace(/\./g, '/');
+