+++ packages/plugin-packages/mashroom-messaging-external-provider-amqp/type-definitions/index.d.ts
@@ -1,0 +1,8 @@
+// @flow
+
+import {MashroomMessagingExternalProvider} from '@mashroom/mashroom-messaging/type-definitions';
+
+export interface MashroomMessagingExternalProviderAMQP extends MashroomMessagingExternalProvider {
+    subscribeToInternalTopic(): void;
+    unsubscribeFromInternalTopic(): void;
+}