+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [unreleased]
 
+ * Added support for messaging via AMQP (Advanced Messaging Queuing) protocol, supported by almost all message brokers
+   (RabbitMQ, Qpid, ActiveMQ, Artemis, Azure Service Bus, ...)
  * Added Remote Portal App registry that automatically scans Kubernetes namespaces for apps
  * Tabify App: The tab buttons have now a new attribute (*data-app-ref*) that contains the id of the corresponding app wrapper div.
    This is useful for end-2-end tests to determine if an app is visible.